requirements.txt
# Core Dependencies
python-okx>=0.3.0          # OKX exchange API
pandas>=2.0.0              # Data manipulation
numpy>=1.24.0              # Numerical computing

# Machine Learning
xgboost>=2.0.0             # Gradient boosting for price prediction
stable-baselines3>=2.2.0   # Reinforcement learning (PPO)
gymnasium>=0.29.0          # RL environments
scikit-learn>=1.3.0        # ML utilities

# Technical Analysis
ta>=0.11.0                 # 50+ technical indicators

# Security
cryptography>=41.0.0       # AES encryption for API keys
python-dotenv>=1.0.0       # Environment variable loading

# Scheduling & Utilities
schedule>=1.2.0            # Task scheduling
requests>=2.31.0           # HTTP requests

# Web Dashboard
flask>=3.0.0               # Web dashboard for remote monitoring
psutil>=5.9.0              # System health monitoring
